Widely expanded the try catch cases and fixed multiple small issues

+++ BFME1/Program.cs
@@ -19,56 +19,49 @@ namespace PatchLauncher
         {
             ApplicationConfiguration.Initialize();
 
-            using Mutex mutex = new(false, Process.GetCurrentProcess().ProcessName);
-            if (!mutex.WaitOne(0, false))
+            try
             {
-                MessageBox.Show(Strings.Warning_AlreadyRunning, Strings.Warning_AlreadyRunningTitle);
-                return;
-            }
+                using Mutex mutex = new(false, Process.GetCurrentProcess().ProcessName);
+                if (!mutex.WaitOne(0, false))
+                {
+                    MessageBox.Show(Strings.Warning_AlreadyRunning, Strings.Warning_AlreadyRunningTitle);
+                    return;
+                }
 
-            if (!Directory.Exists(ConstStrings.C_LOGFOLDER_NAME))
-            {
-                try
+                if (!Directory.Exists(ConstStrings.C_LOGFOLDER_NAME))
                 {
                     Directory.CreateDirectory(ConstStrings.C_LOGFOLDER_NAME);
                 }
-                catch (Exception ex)
-                {
-                    using StreamWriter file = new(Path.Combine(ConstStrings.C_LOGFOLDER_NAME, ConstStrings.C_ERRORLOGGING_FILE), append: true);
-                    file.WriteLineAsync(ex.Message);
-                }
-            }
 
-            if (Directory.Exists(ConstStrings.C_WEBVIEW2CACHEFOLDER_NAME))
-            {
-                try
+                if (Directory.Exists(ConstStrings.C_WEBVIEW2CACHEFOLDER_NAME))
                 {
                     Directory.Delete(ConstStrings.C_WEBVIEW2CACHEFOLDER_NAME, true);
                 }
-                catch (Exception ex)
+
+                string configPath = ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.PerUserRoamingAndLocal).FilePath;
+                if (!File.Exists(configPath))
                 {
-                    File.AppendAllText(Path.Combine(ConstStrings.C_LOGFOLDER_NAME, ConstStrings.C_LOGFOLDER_NAME, "webView2_error.log"), ex.Message);
+                    Settings.Default.Upgrade();
+                    Settings.Default.Reload();
+                    Settings.Default.Save();
+                }
+
+                switch (Settings.Default.Language)
+                {
+                    case 0:
+                        Thread.CurrentThread.CurrentUICulture = new System.Globalization.CultureInfo("en");
+                        break;
+                    case 1:
+                        Thread.CurrentThread.CurrentUICulture = new System.Globalization.CultureInfo("de");
+                        break;
+                    default:
+                        break;
                 }
             }
-
-            string configPath = ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.PerUserRoamingAndLocal).FilePath;
-            if (!File.Exists(configPath))
+            catch (Exception ex)
             {
-                Settings.Default.Upgrade();
-                Settings.Default.Reload();
-                Settings.Default.Save();
-            }
-
-            switch (Settings.Default.Language)
-            {
-                case 0:
-                    Thread.CurrentThread.CurrentUICulture = new System.Globalization.CultureInfo("en");
-                    break;
-                case 1:
-                    Thread.CurrentThread.CurrentUICulture = new System.Globalization.CultureInfo("de");
-                    break;
-                default:
-                    break;
+                using StreamWriter file = new(Path.Combine(ConstStrings.C_LOGFOLDER_NAME, ConstStrings.C_ERRORLOGGING_FILE), append: true);
+                file.WriteLineAsync(ConstStrings.LogTime + ConstStrings.LogTime + ex.ToString());
             }
 
             Application.Run(new WinFormsMainGUI());
